+++ packages/eth-json-rpc-middleware/test/util/helpers.ts
@@ -1,8 +1,30 @@
-import type { SafeEventEmitterProvider } from '@metamask/eth-json-rpc-provider';
-import type { JsonRpcMiddleware } from '@metamask/json-rpc-engine';
+import { PollingBlockTracker } from '@metamask/eth-block-tracker';
+import {
+  providerFromEngine,
+  type SafeEventEmitterProvider,
+} from '@metamask/eth-json-rpc-provider';
+import {
+  JsonRpcEngine,
+  type JsonRpcMiddleware,
+} from '@metamask/json-rpc-engine';
 import type { Json, JsonRpcParams, JsonRpcRequest } from '@metamask/utils';
 import { klona } from 'klona/full';
 import { isDeepStrictEqual } from 'util';
+
+export const createRequest = <
+  Input extends Partial<JsonRpcRequest<Json[]>>,
+  Output extends Input & JsonRpcRequest<Json[]>,
+>(
+  request: Input,
+): Output => {
+  return {
+    jsonrpc: '2.0',
+    id: request.id ?? '1',
+    method: request.method ?? 'test_request',
+    // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+    params: request.params === undefined ? [] : request.params,
+  } as Output;
+};
 
 /**
  * An object that can be used to assign a canned result to a request made via
@@ -47,7 +69,7 @@ export type ProviderRequestStub<
  * @template Result - The type that represents the result.
  * @returns The created middleware, as a mock function.
  */
-export function buildFinalMiddlewareWithDefaultResult<
+export function createFinalMiddlewareWithDefaultResult<
   Params extends JsonRpcParams,
   Result extends Json,
 >(): JsonRpcMiddleware<Params, Result | 'default result'> {
@@ -67,12 +89,53 @@ export function buildFinalMiddlewareWithDefaultResult<
 }
 
 /**
+ * Creates a provider and block tracker. The provider is the block tracker's
+ * provider.
+ *
+ * @returns The provider and block tracker.
+ */
+export function createProviderAndBlockTracker() {
+  const engine = new JsonRpcEngine();
+  const provider = providerFromEngine(engine);
+
+  const blockTracker = new PollingBlockTracker({
+    provider,
+  });
+
+  return { provider, blockTracker };
+}
+
+/**
+ * Creates a JSON-RPC engine with the middleware under test and any
+ * additional middleware. If no other middleware is provided, a final middleware
+ * that returns a default result is added.
+ *
+ * @param middlewareUnderTest - The middleware under test.
+ * @param otherMiddleware - Any additional middleware.
+ * @returns The created engine.
+ */
+export function createEngine(
+  middlewareUnderTest: JsonRpcMiddleware<any, any>,
+  ...otherMiddleware: JsonRpcMiddleware<any, any>[]
+): JsonRpcEngine {
+  const engine = new JsonRpcEngine();
+  engine.push(middlewareUnderTest);
+  if (otherMiddleware.length === 0) {
+    otherMiddleware.push(createFinalMiddlewareWithDefaultResult());
+  }
+  for (const middleware of otherMiddleware) {
+    engine.push(middleware);
+  }
+  return engine;
+}
+
+/**
  * Creates a middleware function that just ends the request, but is also a Jest
  * mock function so that you can make assertions on it.
  *
  * @returns The created middleware, as a mock function.
  */
-export function buildSimpleFinalMiddleware() {
+export function createSimpleFinalMiddleware() {
   return jest.fn((_req, _res, _next, end) => {
     end();
   });
@@ -82,14 +145,14 @@ export function buildSimpleFinalMiddleware() {
  * Some JSON-RPC endpoints take a "block" param (example: `eth_blockNumber`)
  * which can optionally be left out. Additionally, the endpoint may support some
  * number of arguments, although the "block" param will always be last, even if
- * it is optional. Given this, this function builds a `params` array for such an
+ * it is optional. Given this, this function creates a `params` array for such an
  * endpoint with the given "block" param added at the end.
  *
  * @param blockParamIndex - The index within the `params` array to add the "block" param.
  * @param blockParam - The desired "block" param to add.
  * @returns The mock params.
  */
-export function buildMockParamsWithBlockParamAt(
+export function createMockParamsWithBlockParamAt(
   blockParamIndex: number,
   blockParam: string,
 ): string[] {
@@ -107,7 +170,7 @@ export function buildMockParamsWithBlockParamAt(
  * Some JSON-RPC endpoints take a "block" param (example: `eth_blockNumber`)
  * which can optionally be left out. Additionally, the endpoint may support some
  * number of arguments, although the "block" param will always be last, even if
- * it is optional. Given this, this function builds a mock `params` array for
+ * it is optional. Given this, this function creates a mock `params` array for
  * such an endpoint, filling it with arbitrary values, but with the "block"
  * param missing.
  *
@@ -115,7 +178,7 @@ export function buildMockParamsWithBlockParamAt(
  * would* appear.
  * @returns The mock params.
  */
-export function buildMockParamsWithoutBlockParamAt(
+export function createMockParamsWithoutBlockParamAt(
   blockParamIndex: number,
 ): string[] {
   const params = [];
@@ -128,14 +191,14 @@ export function buildMockParamsWithoutBlockParamAt(
 }
 
 /**
- * Builds a canned result for a `eth_blockNumber` request made to
+ * Creates a canned result for a `eth_blockNumber` request made to
  * `provider.request` such that the result will return the given block
  * number. Intended to be used in conjunction with `stubProviderRequests`.
  *
  * @param blockNumber - The block number (default: '0x0').
  * @returns The request/result pair.
  */
-export function buildStubForBlockNumberRequest(
+export function createStubForBlockNumberRequest(
   blockNumber = '0x0',
 ): ProviderRequestStub<JsonRpcParams, Json> {
   return {
@@ -148,7 +211,7 @@ export function buildStubForBlockNumberRequest(
 }
 
 /**
- * Builds a canned result for a request made to `provider.request`. Intended
+ * Creates a canned result for a request made to `provider.request`. Intended
  * to be used in conjunction with `stubProviderRequests`. Although not strictly
  * necessary, it helps to assign a proper type to a request/result pair.
  *
@@ -157,7 +220,7 @@ export function buildStubForBlockNumberRequest(
  * @param requestStub - The request/result pair.
  * @returns The request/result pair, properly typed.
  */
-export function buildStubForGenericRequest<
+export function createStubForGenericRequest<
   Params extends JsonRpcParams,
   Result extends Json,
 >(requestStub: ProviderRequestStub<Params, Result>) {